backend/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
culqi
coreapi==2.3.3
Pillow==10.0.0
